backend/indexer/requirements.txt
psycopg2-binary
opensearch-py
